% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute.R
\name{absolute}
\alias{absolute}
\title{Absolute}
\usage{
absolute(x, val, a, b)
}
\arguments{
\item{x}{// Domain vector}

\item{val}{// Scalar to determine slope}

\item{a}{// Lower bound scalar}

\item{b}{// Upper bound scalar}
}
\value{
returns a vector y
}
\description{
This function takes in a domain vector x, a slope scalar val, a lower bound a, and an upper bound b and generates
a vector y equal to (val * |x|) when a a <= x <= b, and zero otherwise.
}
\examples{
# Creating vector y equal to |x| for -1 <= x <= 1, and zero otherwise over the domain (-3, 3)

x <- seq(-3, 3, by = .002)
y <- absolute(x, 1, -1, 1)
}
