% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{convolution}
\alias{convolution}
\title{Convolution}
\usage{
convolution(f, g, width)
}
\arguments{
\item{f}{// Function vector corresponding to f(x)}

\item{g}{// Function vector corresponding to g(x - t)}

\item{width}{// The length of the interval over which the convolution is being calculated}
}
\value{
returns a scalar equal to the average value of f(x)*g(x - t)
}
\description{
Convolution
}
\examples{
# Generating a function, a kernel equal to the function shifted 50 indices to the right, and discretely calculating the value of the area under function*kernel

x <- seq(-3, 3, by = .002)

function <- absolute(x, 1, -1, 1)
kernel <- ashift(function, 50)

convolution_value <- convolution(function, kernel, length(function))
}
