% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mp4.R
\name{generate_mp4}
\alias{generate_mp4}
\title{Generate .mp4}
\usage{
generate_mp4(
  input_function = NULL,
  bound = 2,
  step_size = 0.002,
  freq = 4,
  fps = 30,
  nIter = 1
)
}
\arguments{
\item{input_function}{// Vector of function values for use as base function and kernel}

\item{bound}{// Radius of the x-axis for the plot}

\item{step_size}{// Distance between each value in the input function}

\item{freq}{// Dictates the number of indices to skip before each frame is generated}

\item{fps}{// Frame rate for the .mp4 file}

\item{nIter}{// Number of convolutions to generate in the .mp4 file}
}
\description{
This function Generates a .mp4 file of a visual representation of the convolution of the input function with itself. If nIter is greater than 1,
then at the end of each iteration, the function is replaced with the convolution, with the kernel remaining constant. Frequency and fps
determine the number of total plot frames and the frame rate of the .mp4, respectively. Bound and step size indicate the radius of the plot
and the distance between values over which the input function is calculated.
}
\examples{
# Generating a single loop .mp4 using a uniform distribution on (-.5,.5), with step size equal to 1/500,
# 4 indices skipped, and a frame rate of 30.

x <- seq(-2, 2, by = .002)

y <- unif(x, -.5, .5)

generate_mp4(input_function = y, bound = 2, step_size = .002, freq = 4, fps = 30, nIter = 1)
}
