% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_frame.R
\name{last_frame}
\alias{last_frame}
\title{Last Frame}
\usage{
last_frame(input_function = NULL, bound = 2, step_size = 0.002, freq = 4)
}
\arguments{
\item{input_function}{// Vector of function values for use as base function and kernel}

\item{bound}{// Radius of the x-axis for the plot}

\item{step_size}{// Distance between each value in the input function}

\item{freq}{// Dictates the number of indices to skip before each frame is generated}
}
\description{
This function generates an image file of the potential last frame (in the first loop) in the .mp4 file that would be generated by generate_mp4().
It takes in an input function, the bound equal to the radius of the x-axis, a step size that indicates the distance between domain values, and a frequency that
determines the number of indices skipped between plot frames.
}
\examples{
# Checking the last frame for a uniform distribution on (-.5, 5) with a plot radius of 2.

x <- seq(-2, 2, by = .002)

y <- unif(x, -.5, .5)

last_frame(input_function = y, bound = 2, step_size = .002, freq = 4)

}
