% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Linear}
\usage{
linear(x, slope, intercept, a, b)
}
\arguments{
\item{x}{// Domain vector}

\item{slope}{// Scalar to indicate slope}

\item{intercept}{// Scalar to indicate x-intercept}

\item{a}{// Lower bound scalar}

\item{b}{// Upper bound scalar}
}
\value{
returns a vector y
}
\description{
This function takes in a domain vector x, and scalars to indicate slope, intercept, lower bound a, and upper bound b, and
returns a vector y equal to (slope * x + intercept) for a <= x <= b, and zero otherwise, defined over domain x.
}
\examples{
# Generating y = -.5x + .5 for -1 <= x <= 1 over the domain (-3, 3)

x <- seq(-3, 3, by = .002)
y <- linear(x, -.5, -1, 1)

}
